chore: move all logic to the same place
